Dockerfile-hermes: Add Dockerfile for eid-hermes project
Add a Dockerfile that pulls the Eideticom eid-hermes-qemu repo and
builds the x86 system emulator. It uses a rootfs for this VM that is
generated using gen-images (using cloud-init) and spins the whole
thing up.

Note that the cloud-init networking should allow for instance SSH into
the VM via port 2222. The username and password for the main account
on the VM is 'hermes' and 'zeus'. Note we add a .dockerignore to avoid
pulling the large images into context. We also include a
docker-compose.yml for easier management and a README.md to summarize
this approach.