[temp] Add dumps and logs from AWS F1 FPGA
We want to make hermes like the AWS F1 FPGAs, so we need to see how it
behaves and is configured.

I loaded the cl_dram_dma AFI[1] and probed BARs 0, 1 and 2 to see their
current values. We add them here so that we can compare with Hermes.
Probably only BAR2 needs to match.

I also compiled the XDMA driver in debug mode and ran a DMA (using part
of the program described here[2], up to the pwrite()) and captured the
xdma log on dmesg. I also captured the xdma log when I inserted the
module.

[1]: https://github.com/aws/aws-fpga/tree/master/hdk/cl/examples/cl_dram_dma
[2]: https://github.com/aws/aws-fpga/tree/master/sdk/linux_kernel_drivers/xdma#quick-example

Signed-off-by: Martin Oliveira <martin.oliveira@eideticom.com>