Add libubpf to configure script
uBPF[1] is a userspace eBPF virtual machine, which includes an
assembler, disassembler, interpreter and JIT compiler for x86-64.

Because it is not part of the kernel, it can be used by userspace tools
that need to operate on programs in the eBPF ISA.

We need this to enable the hermes device model.

There is currently an open issue on uBPF[2] that requires the inclusion
of stdbool.h as well as ubpf.h. A PR has been opened[3], but has not yet
been merged into upstream, so we add that header to our configure script.

[1]: https://github.com/iovisor/ubpf
[2]: https://github.com/iovisor/ubpf/issues/38
[3]: https://github.com/iovisor/ubpf/pull/39

Signed-off-by: Martin Oliveira <martin.oliveira@eideticom.com>