Add Hermes BAR 2
BAR 2 contains the registers used by the XDMA driver. The layout of
this register is decribed on on Xilinx's manual PG195[1]. Note that that
document uses BAR1, but because on AWS F1 FPGAs it is mapped to BAR 2,
we use BAR 2 here as well.

This commits adds all of the registers, but does not perform DMAs yet.

[1]: https://www.xilinx.com/support/documentation/ip_documentation/xdma/v4_1/pg195-pcie-dma.pdf

Signed-off-by: Martin Oliveira <martin.oliveira@eideticom.com>