Add Hermes BAR 2
BAR 2 contains the registers used by the XDMA driver. The layout of
this register is decribed on on Xilinx's manual PG195[1]. Note that that
document uses BAR1, but because on AWS F1 FPGAs it is mapped to BAR 2,
we use BAR 2 here as well.

This commits the registers, but does not perform DMAs yet. We only add
registers that are effectively used by the xdma driver[2].

We initialize registers to match AWS F1.

[1]: https://www.xilinx.com/support/documentation/ip_documentation/xdma/v4_1/pg195-pcie-dma.pdf
[2]: https://github.com/aws/aws-fpga/tree/master/sdk/linux_kernel_drivers/xdma

Signed-off-by: Martin Oliveira <martin.oliveira@eideticom.com>