Compute number of descriptors in next DMA block correctly
DMA descriptors are processed in blocks, and the last descriptor of a
block points to the address and number of *adjacent* descriptors in the
next block (Nxt_adj and Nxt_addr on Table 5 of PG195[1]).

However, we were not computing the number of descriptors in subsequent
blocks properly, because we weren't adding 1 to account for the first
descriptor (it is valid to have nxt_adj == 0).

This has been tested with a 32MB XDMA transfer.

Fixes: 9d38cf97af58 ("Implement DMA on Hermes")

[1]: https://www.xilinx.com/support/documentation/ip_documentation/xdma/v4_1/pg195-pcie-dma.pdf

Signed-off-by: Martin Oliveira <martin.oliveira@eideticom.com>